requirements.txt
docling>=2.0.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-docx>=1.1.0
docx2python>=2.0.0
watchdog>=3.0.0
tiktoken>=0.5.1
